backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pdfminer.six>=20231228
python-docx>=1.1.0
spacy>=3.7.4
pandas>=2.2.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
